v1.1.0

+++ common/components/base/AccessTokenSystem.php
@@ -5,6 +5,7 @@ namespace common\components\base;
 use common\helpers\ArrayHelper;
 use api\models\User;
 use Yii;
+use yii\web\UnauthorizedHttpException;
 
 /**
  * Class AccessTokenSystem
@@ -26,16 +27,38 @@ class AccessTokenSystem extends \yii\base\Component
      */
     public function getAccessToken(User $model)
     {
-        !empty($model->access_token) && Yii::$app->cache->delete($this->getCacheKey($model->access_token));
+        if (empty($model->access_token) || !$this->timeValid($model->access_token, intval(Yii::$app->params['user']['accessTokenExpired']))) {
+            !empty($model->access_token) && Yii::$app->cache->delete($this->getCacheKey($model->access_token));
+            $model->access_token = Yii::$app->security->generateRandomString() . '_' . time();
+            $model->refresh_token = Yii::$app->security->generateRandomString() . '_' . time();
+            if (!$model->save(false)) {
+                Yii::$app->logSystem->db($model->errors);
+            }
+            $this->saveToCache($this->getCacheKey($model->access_token), $model);
+        }
+
+        Yii::$app->user->login($model);
+        return $model;
+    }
+
+    public function refreshToken(User $model) {
         $model->access_token = Yii::$app->security->generateRandomString() . '_' . time();
         $model->refresh_token = Yii::$app->security->generateRandomString() . '_' . time();
         if (!$model->save(false)) {
             Yii::$app->logSystem->db($model->errors);
+            return false;
+        }
+        return true;
+    }
+
+    public function timeValid($token, $expire)
+    {
+        $time = intval(substr($token, strrpos($token, '_') + 1));
+        if (($time + $expire) < time()) {
+            return false;
         }
 
-        Yii::$app->user->login($model);
-        $this->saveToCache($this->getCacheKey($model->access_token), $model);
-        return $model;
+        return true;
     }
 
     public function disableAccessToken($accessToken)
@@ -68,6 +91,6 @@ class AccessTokenSystem extends \yii\base\Component
 
     public function getFromCache($token, $type = null)
     {
-        return Yii::$app->cache->get($this->getCacheKey($token)) ?? User::findUserByAccessToken($token, $type);
+        return Yii::$app->cache->get($this->getCacheKey($token)) ?: User::findUserByAccessToken($token, $type);
     }
 }
